pyproject.toml
[tool.poetry]
name = "hotel"
version = "0.1.0"
description = "Simple hotel rooms and bookings API (Django)."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "hotel" }, { include = "booking" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
Django = "^5.2"
djangorestframework = "^3.16"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
pytest = "^8.3"
pytest-django = "^4.9"
pre-commit = "^4.0"

[tool.pytest.ini_options]
django_find_project = true
DJANGO_SETTINGS_MODULE = "hotel.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-q"

[tool.poetry.scripts]
manage = "django.core.management:execute_from_command_line"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
